backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Elijah Khor <elijahkcm@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
uvicorn = "^0.34.0"
fastapi = "^0.115.8"
gunicorn = "^23.0.0"
python-dotenv = "^1.0.1"
python-multipart = "^0.0.20"
langchain-pinecone = "^0.2.2"
langchain-community = "^0.3.17"
pdfplumber = "^0.11.5"
openai = "^1.61.1"
boto3 = "^1.36.16"
langchain-openai = "^0.3.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
